Replace logic for process event to SmartHome class
